task1 only

+++ task1.js
@@ -3,8 +3,8 @@
 "use strict";
 
 var i = process.argv[2],
-    a = parseFloat(process.argv[3]),
-    b = parseFloat(process.argv[4]),
+    a = parseInt(process.argv[3]),
+    b = parseInt(process.argv[4]),
     normal = 0,
     result = "",
     numberTranslation = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 'a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j', 'k', 'l', 'm', 'n', 'o', 'p', 'q', 'r', 's', 't', 'u', 'v', 'w', 'x', 'y', 'z'];
